requirements.txt
requests>=2.31.0
PyPDF2>=3.0.1
python-dotenv>=1.0.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
anthropic>=0.18.0
chromadb>=0.4.0
torch>=1.11.0
sentence-transformers>=2.2.0
pydantic>=2.0.0
aiofiles>=23.0.0
pytest>=7.0.0
